+++ sdk/resourcemanager/storage/armstorage/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 4.0.0 (2026-05-07)
+## 4.0.0 (2026-05-19)
 ### Breaking Changes
 
 - Function `*BlobContainersClient.CreateOrUpdateImmutabilityPolicy` parameter(s) have been changed from `(ctx context.Context, resourceGroupName string, accountName string, containerName string, options *BlobContainersClientCreateOrUpdateImmutabilityPolicyOptions)` to `(ctx context.Context, resourceGroupName string, accountName string, containerName string, parameters ImmutabilityPolicy, options *BlobContainersClientCreateOrUpdateImmutabilityPolicyOptions)`
@@ -40,29 +40,35 @@
 - New function `*ClientFactory.NewConnectorsClient() *ConnectorsClient`
 - New function `*ClientFactory.NewDataSharesClient() *DataSharesClient`
 - New function `*ConnectorAuthProperties.GetConnectorAuthProperties() *ConnectorAuthProperties`
+- New function `*ConnectorAuthPropertiesUpdate.GetConnectorAuthPropertiesUpdate() *ConnectorAuthPropertiesUpdate`
 - New function `*ConnectorConnection.GetConnectorConnection() *ConnectorConnection`
 - New function `*ConnectorSource.GetConnectorSource() *ConnectorSource`
+- New function `*ConnectorSourceUpdate.GetConnectorSourceUpdate() *ConnectorSourceUpdate`
 - New function `NewConnectorsClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*ConnectorsClient, error)`
 - New function `*ConnectorsClient.BeginCreate(ctx context.Context, resourceGroupName string, accountName string, connectorName string, resource Connector, options *ConnectorsClientBeginCreateOptions) (*runtime.Poller[ConnectorsClientCreateResponse], error)`
 - New function `*ConnectorsClient.BeginDelete(ctx context.Context, resourceGroupName string, accountName string, connectorName string, options *ConnectorsClientBeginDeleteOptions) (*runtime.Poller[ConnectorsClientDeleteResponse], error)`
 - New function `*ConnectorsClient.Get(ctx context.Context, resourceGroupName string, accountName string, connectorName string, options *ConnectorsClientGetOptions) (ConnectorsClientGetResponse, error)`
 - New function `*ConnectorsClient.NewListByStorageAccountPager(resourceGroupName string, accountName string, options *ConnectorsClientListByStorageAccountOptions) *runtime.Pager[ConnectorsClientListByStorageAccountResponse]`
 - New function `*ConnectorsClient.BeginTestExistingConnection(ctx context.Context, resourceGroupName string, accountName string, connectorName string, body TestExistingConnectionRequest, options *ConnectorsClientBeginTestExistingConnectionOptions) (*runtime.Poller[ConnectorsClientTestExistingConnectionResponse], error)`
-- New function `*ConnectorsClient.BeginUpdate(ctx context.Context, resourceGroupName string, accountName string, connectorName string, properties Connector, options *ConnectorsClientBeginUpdateOptions) (*runtime.Poller[ConnectorsClientUpdateResponse], error)`
+- New function `*ConnectorsClient.BeginUpdate(ctx context.Context, resourceGroupName string, accountName string, connectorName string, properties ConnectorUpdate, options *ConnectorsClientBeginUpdateOptions) (*runtime.Poller[ConnectorsClientUpdateResponse], error)`
 - New function `*DataShareConnection.GetConnectorConnection() *ConnectorConnection`
 - New function `*DataShareSource.GetConnectorSource() *ConnectorSource`
+- New function `*DataShareSourceUpdate.GetConnectorSourceUpdate() *ConnectorSourceUpdate`
 - New function `NewDataSharesClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*DataSharesClient, error)`
 - New function `*DataSharesClient.BeginCreate(ctx context.Context, resourceGroupName string, accountName string, dataShareName string, resource DataShare, options *DataSharesClientBeginCreateOptions) (*runtime.Poller[DataSharesClientCreateResponse], error)`
 - New function `*DataSharesClient.BeginDelete(ctx context.Context, resourceGroupName string, accountName string, dataShareName string, options *DataSharesClientBeginDeleteOptions) (*runtime.Poller[DataSharesClientDeleteResponse], error)`
 - New function `*DataSharesClient.Get(ctx context.Context, resourceGroupName string, accountName string, dataShareName string, options *DataSharesClientGetOptions) (DataSharesClientGetResponse, error)`
 - New function `*DataSharesClient.NewListByStorageAccountPager(resourceGroupName string, accountName string, options *DataSharesClientListByStorageAccountOptions) *runtime.Pager[DataSharesClientListByStorageAccountResponse]`
-- New function `*DataSharesClient.BeginUpdate(ctx context.Context, resourceGroupName string, accountName string, dataShareName string, properties DataShare, options *DataSharesClientBeginUpdateOptions) (*runtime.Poller[DataSharesClientUpdateResponse], error)`
+- New function `*DataSharesClient.BeginUpdate(ctx context.Context, resourceGroupName string, accountName string, dataShareName string, properties DataShareUpdate, options *DataSharesClientBeginUpdateOptions) (*runtime.Poller[DataSharesClientUpdateResponse], error)`
 - New function `*ManagedIdentityAuthProperties.GetConnectorAuthProperties() *ConnectorAuthProperties`
+- New function `*ManagedIdentityAuthPropertiesUpdate.GetConnectorAuthPropertiesUpdate() *ConnectorAuthPropertiesUpdate`
 - New function `*TaskAssignmentsClient.BeginStopAssignment(ctx context.Context, resourceGroupName string, accountName string, storageTaskAssignmentName string, options *TaskAssignmentsClientBeginStopAssignmentOptions) (*runtime.Poller[TaskAssignmentsClientStopAssignmentResponse], error)`
 - New struct `AccountSharedKeyAccessProperties`
 - New struct `Connector`
 - New struct `ConnectorListResult`
 - New struct `ConnectorProperties`
+- New struct `ConnectorPropertiesUpdate`
+- New struct `ConnectorUpdate`
 - New struct `DataCollaborationPolicyProperties`
 - New struct `DataShare`
 - New struct `DataShareAccessPolicy`
@@ -70,8 +76,12 @@
 - New struct `DataShareConnection`
 - New struct `DataShareListResult`
 - New struct `DataShareProperties`
+- New struct `DataSharePropertiesUpdate`
 - New struct `DataShareSource`
+- New struct `DataShareSourceUpdate`
+- New struct `DataShareUpdate`
 - New struct `ManagedIdentityAuthProperties`
+- New struct `ManagedIdentityAuthPropertiesUpdate`
 - New struct `ObjectReplicationPolicyPropertiesTagsReplication`
 - New struct `ServiceSharedKeyAccessProperties`
 - New struct `StaticWebsite`